+++ opentelemetry-sdk/src/export/logs/mod.rs
@@ -1,10 +1,11 @@
 //! Log exporters
+use crate::logs::LogRecord;
 use crate::Resource;
 use async_trait::async_trait;
 #[cfg(feature = "logs_level_enabled")]
 use opentelemetry::logs::Severity;
 use opentelemetry::{
-    logs::{LogError, LogRecord, LogResult},
+    logs::{LogError, LogResult},
     InstrumentationLibrary,
 };
 use std::fmt::Debug;